+++ include/protocols.h
@@ -908,12 +908,11 @@ ISO 7816-4 Basic interindustry commands. For command APDU's.
 #define HITAG1_HALT                     0x70    // left 4 bits only, followed by 8 bits (dummy) page and 8 bits CRC
 
 // HITAG2 commands
-#define HITAG2_START_AUTH               0x3    // left 5 bits only
-#define HITAG2_HALT                     0x0    // left 5 bits only
-
-#define HITAG2_READ_PAGE                0x3    // page number in bits 5 to 3, page number inverted in bit 0 and following 2 bits
-#define HITAG2_READ_PAGE_INVERTED       0x1    // page number in bits 5 to 3, page number inverted in bit 0 and following 2 bits
-#define HITAG2_WRITE_PAGE               0x2   // page number in bits 5 to 3, page number
+#define HITAG2_START_AUTH           "11000"         // get UID and/or start the authentication process
+#define HITAG2_READ_PAGE            "11"            // read page after auth
+#define HITAG2_READ_PAGE_INVERTED   "01"            // as read page but all bits inverted
+#define HITAG2_WRITE_PAGE           "10"            // write page after auth
+#define HITAG2_HALT                 "00"            // silence currently authenticated tag
 
 
 // HITAG S commands